// src/app/Modules/layout-settings/Model/layoutposition.model.ts
export class LayoutPosition {
    public layoutPositionName: string = '';
    public positionShortName: number;
    public layoutCategoryName:string='';
    public positionId:number;
    responseModel: LayoutPosition[];
    value: LayoutPosition;
    active_status: number;
    keyword: string;
    activeStatus: string;
    pageSize?: number;
    page?: number;
    count?:number
  }